# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Alhumoudi Lawyers – practice management platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    patch:
      operationId: updateUser
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted

  # ─────────────────────────────────────────────
  # CASES
  # ─────────────────────────────────────────────
...
  /cases/{caseId}/documents:
    get:
      operationId: listCaseDocuments
      tags: [documents]
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Documents for case
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
...
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
    patch:
      operationId: updateDocument
      tags: [documents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    delete:
      operationId: deleteDocument
      tags: [documents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
  /documents/{id}/versions:
    get:
      operationId: getDocumentVersions
      tags: [documents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Version history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocumentVersion"

  # ─────────────────────────────────────────────
  # TIME ENTRIES
  # ─────────────────────────────────────────────
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      parameters:
        - name: caseId
          in: query
          schema:
            type: number
        - name: assigneeId
          in: query
          schema:
            type: number
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    patch:
      operationId: updateTask
      tags: [tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted

  # ─────────────────────────────────────────────
  # CONTRACTS
  # ─────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        email:
          type: string
        role:
          type: string
          description: "admin | lawyer | support"
        barNumber:
          type: ["string", "null"]
        specialization:
          type: ["string", "null"]
        billableRate:
          type: ["number", "null"]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
    UserInput:
      type: object
      required: [name, email, role]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        role:
          type: string
        barNumber:
          type: string
        specialization:
          type: string
        billableRate:
          type: number
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
        barNumber:
          type: string
        specialization:
          type: string
        billableRate:
          type: number
        active:
          type: boolean

    # ── Cases ──────────────────────────────────
...
    Document:
      type: object
      required: [id, caseId, title, fileType, status, version, uploadedBy, createdAt]
      properties:
        id:
          type: number
        caseId:
          type: number
        caseName:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        fileType:
          type: string
        filePath:
          type: ["string", "null"]
        status:
          type: string
          description: "draft | final | archived"
        version:
          type: number
        uploadedBy:
          type: string
        uploadedById:
          type: ["integer", "null"]
        tags:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DocumentInput:
      type: object
      required: [title, fileType]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        fileType:
          type: string
        filePath:
          type: string
        status:
          type: string
        uploadedById:
          type: number
        tags:
          type: string
    DocumentUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
        tags:
          type: string
    DocumentVersion:
      type: object
      required: [id, documentId, version, editedBy, editedAt, changeNote]
      properties:
        id:
          type: number
        documentId:
          type: number
        version:
          type: number
        editedBy:
          type: string
        editedAt:
          type: string
          format: date-time
        changeNote:
          type: string

    # ── Time Entries ────────────────────────────
...
    Task:
      type: object
      required: [id, title, status, priority, createdAt]
      properties:
        id:
          type: number
        title:
          type: string
        description:
          type: ["string", "null"]
        caseId:
          type: ["integer", "null"]
        caseName:
          type: ["string", "null"]
        assigneeId:
          type: ["integer", "null"]
        assigneeName:
          type: ["string", "null"]
        createdById:
          type: ["integer", "null"]
        status:
          type: string
          description: "pending | in_progress | done | cancelled"
        priority:
          type: string
          description: "low | medium | high | urgent"
        dueDate:
          type: ["string", "null"]
          format: date
        completedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    TaskInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        caseId:
          type: number
        assigneeId:
          type: number
        createdById:
          type: number
        status:
          type: string
        priority:
          type: string
        dueDate:
          type: string
          format: date
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        caseId:
          type: number
        assigneeId:
          type: number
        status:
          type: string
        priority:
          type: string
        dueDate:
          type: string
          format: date

    # ── Contracts ──────────────────────────────
...