# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Alhumoudi Lawyers – practice management platform API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  name: string;
  email: string;
  /** admin | lawyer | support */
  role: string;
  /** @nullable */
  barNumber?: string | null;
  /** @nullable */
  specialization?: string | null;
  /** @nullable */
  billableRate?: number | null;
  active?: boolean;
  createdAt: Date;
}