# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Alhumoudi Lawyers – practice management platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  name?: string;
  email?: string;
  role?: string;
  barNumber?: string;
  specialization?: string;
  billableRate?: number;
  active?: boolean;
}